Implemented HIR and first-pass resolution (#5)
We needed to generate a HIR representation and move to a flat node
structure in order to be able to efficiently look up nodes that have
been resolved to each other. The next step is to use this HIR to infer
and check types. This may produce a MIR representation that will be used
for optimization and codegen.